requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
python-multipart==0.0.20

# AI / Agents
anthropic==0.43.0
crewai==0.95.0
langchain==0.3.14

# Speech
openai-whisper==20240930

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.2.1
alembic==1.14.0

# Communication
twilio==9.4.0
firebase-admin==6.6.0

# HTTP / Async
httpx==0.28.1
aiohttp==3.11.11
websockets==14.1

# Location
googlemaps==4.10.0

# Utilities
structlog==24.4.0
tenacity==9.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1
